requirements.txt
Django
djangorestframework
gunicorn
whitenoise
dj-database-url
psycopg2-binary
mysqlclient